scm/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.0.0
databricks-sdk>=0.20.0
python-dateutil>=2.8.0
